Initial test of refactored event visualizer
